Refractored code so the main controller no longer holds most of the logic
Most of the functionality is now in factories.js which creates factories
for weather, time, display, and city object creation.
The structure of the app slightly changed with the introduction of the city
object, and as an added benefit, the app no longer needs to fetch data
when switching between cities.
Minor syntax and style issues were also fixed in this version as specfied
to do so following the first code commit.